Upload: Move utility classes out of cli folder
